% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{beta_C}
\alias{beta_C}
\title{Calculate beta_C}
\usage{
beta_C(x, C, extrapolation = T, interrupt = T)
}
\arguments{
\item{x}{a site by species matrix}

\item{C}{target coverage. value between 0 and 1.}

\item{extrapolation}{logical. should extrapolation be used?}

\item{interrupt}{logical. SHould the function throw an error when C exceeds the maximum recommendable coverage?}
}
\value{
a numeric value
}
\description{
Beta_C uses coverage-based rarefaction to standardize beta-diversity. Calculates the ratio between 
gamma and alpha scale IBR curve for a target gamma-scale sample coverage (i.e. a measure of sample completeness).
}
\examples{
\donttest{
library(vegan)
data(BCI)

# What is beta_C for a coverage value of 60\%?
beta_C(BCI,C = 0.6)
}
}
