% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{C_target}
\alias{C_target}
\title{Calculate the recommended maximum coverage value for the computation of beta_C}
\usage{
C_target(x, factor = 2)
}
\arguments{
\item{x}{a site by specie matrix}

\item{factor}{numeric. how far do you want to extrapolate?}
}
\value{
numeric value
}
\description{
Returns the estimated gamma-scale coverage that corresponds to the largest allowable sample size 
(i.e. the smallest observed sample size at the alpha scale multiplied by an extrapolation factor).
The default (factor = 2) allows for extrapolation up to 2 times the observed sample size of the 
smallest alpha sample. For factor= 1, only interpolation is applied.
Its not recommendable to use factors larger than 2.
}
\examples{
\donttest{
library(vegan)
data(BCI)

# What is the largest possible C that I can use to calculate beta_C for my site by species matrix?
C_target(BCI)
}

}
