% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{calc_beta_div}
\alias{calc_beta_div}
\title{Calculate beta diversity from sites by species table.}
\usage{
calc_beta_div(x, index, effort = NA, coverage = TRUE, ...)
}
\arguments{
\item{x}{Sites by species table with species abundances}
}
\description{
Calculate beta diversity from sites by species table.
}
\examples{
data(inv_comm)
beta_metrics = calc_beta_div(inv_comm, 'S_n', effort = c(5, 10))
beta_metrics
}
