% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{calc_chao1}
\alias{calc_chao1}
\title{Estimation of species richness}
\usage{
calc_chao1(x)
}
\arguments{
\item{x}{a vector of species abundances}
}
\value{
a vector of observed species richness and a species richness estimate
}
\description{
\code{calc_chao1}: estimation of species richness based on the methods
proposed in Chao (1984, 1987).
}
\details{
This function is a trimmed version of \code{iNext::ChaoRichess} found at 
\url{https://github.com/JohnsonHsieh/iNEXT}. T. C. Hsieh, K. H. Ma and Anne
Chao are the orginal authors of the \code{iNEXT} package.
}
\examples{
data(inv_comm)
calc_chao1(inv_comm)
}
\references{
Chao, A. (1984) Nonparametric estimation of the number of classes in a
population. Scandinavian Journal of Statistics, 11, 265-270.
Chao, A. (1987) Estimating the population size for capture-recapture data with
unequal catchability. Biometrics, 43, 783-791.
}
