% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{calc_comm_div}
\alias{calc_comm_div}
\title{Calculate biodiversity statistics from sites by species table.}
\usage{
calc_comm_div(
  abund_mat,
  index,
  effort = NA,
  extrapolate = TRUE,
  return_NA = FALSE,
  rare_thres = 0.05,
  scales = c("alpha", "gamma", "beta"),
  replace = FALSE,
  C_target_gamma = NA,
  ...
)
}
\arguments{
\item{abund_mat}{Abundance based site-by-species table. Species as
columns}

\item{index}{The calculated biodiversity indices. The options are
\itemize{
   \item \code{N} ... Number of individuals (total abundance)
   \item \code{S} ... Number of species
   \item \code{S_n} ... Rarefied or extrapolated number of species for n individuals
   \item \code{S_C} ... Rarefied or extrapolated number of species for C coverage
   \item \code{S_asymp} ... Estimated asymptotic species richness
   \item \code{f_0} ... Estimated number of undetected species 
   \item \code{pct_rare} ... The percent of species with abundances below \code{rare_thres}
   \item \code{PIE} ... Hurlbert's PIE (Probability of Interspecific Encounter)
   \item \code{S_PIE} ... Effective number of species based on PIE
}

See the documentation of \code{\link{get_mob_stats}} for further details on the
biodiversity indices.}

\item{effort}{The standardized number of individuals used for the 
calculation of rarefied species richness. This can a be
single integer or a vector of integers.}

\item{extrapolate}{Boolean which specifies if richness should be
extrapolated when \code{effort_samples} is larger than the number of
individuals using the chao1 method. Defaults to TRUE.}

\item{return_NA}{Boolean defaults to FALSE in which the rarefaction function
returns the observed S when \code{effort} is larger than the number of
individuals. If set to TRUE then NA is returned. Note that this argument
is only relevant when \code{extrapolate = FALSE}.}

\item{rare_thres}{The threshold that determines how pct_rare is computed.
It can range from (0, 1] and defaults to 0.05 which specifies that any 
species with less than or equal to 5% of the total abundance in a sample is
considered rare. It can also be specified as "N/S" which results in using
average abundance as the threshold which McGill (2011) found to have the 
best small sample behavior.}

\item{scales}{The scales to compute the diversity indices for: 
\itemize{
    \item \code{alpha} ... for each row of the site x species community matrix
    \item \code{gamma} ... for the entire site x species community matrix
    \item \code{beta} ... the ratio of diversity at the \code{gamma} and 
                           \code{alpha} scales.  
} Defaults to all three scales: \code{c('alpha', 'gamma', 'beta')}}

\item{replace}{Used for \code{PIE} and \code{SPIE}.  If TRUE, sampling with
replacement is used. Otherwise, sampling without replacement (default).}

\item{C_target_gamma}{When computing coverage based richness (\code{S_C}) then 
this argument can be used to specify the coverage to be used for the gamma scale
richness estimate. This defaults to \code{NA} in which case the target cover
is computed by \code{\link{calc_C_target}} (i.e., the largest allowable sample
size).}

\item{...}{
Optional arguments to \code{FUN} and also to underlying functions (e.g.
\code{\link{parLapply}} and \code{\link{mclapply}} when \code{cl}
is not \code{NULL}).
}
}
\value{
A \code{data.frame} with four columns:
\itemize{
   \item \code{group} ... Group label for sites
   \item \code{index} ... Name of the biodiversity index
   \item \code{effort} ... Sampling effort for rarefied richness 
   (NA for the other indices)
   \item \code{gamma_coverage} ... The coverage value for that particular 
   effort value on the gamma scale rarefaction curve. Will be \code{NA} unless
   coverage based richness (\code{S_C}) and/or beta diveristy is computed.  
   \item \code{value} ... Value of the biodiversity index
}
}
\description{
Calculate biodiversity statistics from sites by species table.
}
\details{
This function is primarily intended as auxiliary function used in
\code{\link{get_mob_stats}}, but can be also used directly for data exploration.
}
\examples{
data(tank_comm)
div_metrics <- calc_comm_div(tank_comm, 'S_n', effort = c(5, 10))
div_metrics
div_metrics <- calc_comm_div(tank_comm, 'S_C', C_target_gamma = 0.75)
div_metrics
}
\references{
McGill, B. J. 2011. Species abundance distributions. Pages 105-122 Biological
Diversity: Frontiers in Measurement and Assessment, eds. A.E. Magurran and
B.J. McGill.
}
\author{
Felix May and Dan McGlinn
}
