% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{calc_comm_div_ci}
\alias{calc_comm_div_ci}
\title{Compute non-parametric confidence intervals across diversity indices.}
\usage{
calc_comm_div_ci(
  samples,
  cent_stat = "median",
  ci = c(0.025, 0.975),
  index,
  effort = NA,
  extrapolate = TRUE,
  return_NA = FALSE,
  rare_thres = 0.05,
  scales = c("alpha", "gamma", "beta"),
  PIE_replace = FALSE,
  C_target_gamma = NA,
  ...
)
}
\arguments{
\item{samples}{a list of community matrices (i.e., the output of \code{get_samples})}

\item{cent_stat}{a string that is either 'mean' or 'median' which specifies the measure
of central tendency. Defaults to 'median'.}

\item{ci}{a numeric vector of two numbers specifying the lower and upper
quantiles of the distribution to return. The default is to report
the 0.025 and 0.975 quantiles in other words a 95 percent interval.}

\item{index}{The calculated biodiversity indices. The options are
\itemize{
   \item \code{N} ... Number of individuals (total abundance)
   \item \code{S} ... Number of species
   \item \code{S_n} ... Rarefied or extrapolated number of species for n individuals
   \item \code{S_C} ... Estimate species richness of a given level of coverage by \code{C_target_gamma}
   \item \code{S_asymp} ... Estimated asymptotic species richness
   \item \code{f_0} ... Estimated number of undetected species
   \item \code{pct_rare} ... The percent of rare species as defined by \code{rare_thres}
   \item \code{PIE} ... Hurlbert's PIE (Probability of Interspecific Encounter)
   \item \code{S_PIE} ... Effective number of species based on PIE

}
  See \emph{Details} for additional information on the biodiversity
  statistics.}

\item{effort}{The standardized number of individuals used for the calculation
of rarefied species richness. This must be a single integer.}

\item{extrapolate}{Boolean which specifies if richness should be extrapolated
when effort is larger than the number of individuals using the chao1
method (Chao 1984, 1987). Defaults to TRUE.}

\item{return_NA}{Boolean in which the rarefaction function returns the
observed S when \code{effort} is larger than the number of individuals. If
set to TRUE then NA is returned. Note that this argument is only relevant
when \code{extrapolate = FALSE}.}

\item{rare_thres}{The threshold that determines how the metric
\code{pct_rare} is computed. It can range from (0, 1] and defaults to 0.05
which specifies that any species with less than or equal to 5% of the total
abundance in a sample is considered rare. It can also be specified as "N/S"
which results in using average abundance as the threshold which McGill
(2011) found to have the best small sample behavior.}

\item{scales}{The scales to compute the diversity indices for:
\itemize{
    \item \code{alpha} ... for each row of the site x species community matrix
    \item \code{gamma} ... for the entire site x species community matrix
    \item \code{beta} ... the ratio of diversity at the \code{gamma} and
                           \code{alpha} scales.
} Defaults to all three scales: \code{c('alpha', 'gamma', 'beta')}}

\item{PIE_replace}{Used for \code{PIE} and \code{SPIE}.  If TRUE, sampling with
replacement is used. Otherwise, sampling without replacement (default).}

\item{C_target_gamma}{When computing coverage based richness (\code{S_C})
then this argument can be used to specify the coverage to be used for the
gamma scale richness estimate. This defaults to \code{NA} in which case the
target cover is computed by \code{\link{calc_C_target}} (i.e., the largest
allowable sample size).}

\item{...}{additional arguments that can be passed to \code{\link{calc_div}}}
}
\value{
a data.frame that has the lower, middle, and upper quantiles of the 
 sampled distribution of each diversity index.
}
\description{
This function take a list of community matrices and returns the central tendency
and the confidence interval for each diversity index of interest across that
list of communities.
}
\details{
The measure of central tendency can be the median (default) or mean, and the
range of the confidence interval can be specified but defaults to a 95% 
confidence interval.
}
\examples{
data(tank_comm)
samples <- get_samples(tank_comm, algo = 'loo')
calc_comm_div_ci(samples, index = 'S_PIE')
samples <- get_samples(tank_comm, algo = 'boot', n_boot = 20)
calc_comm_div_ci(samples, index = 'S_PIE')
# compute ci for average diversity (rather than median)
calc_comm_div_ci(samples, index = 'S_PIE', cent_stat = 'avg')
}
\seealso{
\code{\link{get_samples}} for generating samples, and \code{\link{calc_comm_div}}
  for the calculation of diversity indices.
}
