% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{calc_PIE}
\alias{calc_PIE}
\title{Calculate probability of interspecific encounter (PIE)}
\usage{
calc_PIE(x)
}
\arguments{
\item{x}{community data, a matrix-like object or a vector which should either
represent a species-abundance distribution or a plot (rows) by species
(columns) matrix.}
}
\description{
PIE is also known as Simpson's evenness index and this function is 
a reduced form of the function vegan::diversity(). Jari Oksanen and Bob O'Hara
are the original authors of the function vegan::diversity()
}
\examples{
data(inv_comm)
calc_PIE(inv_comm)
}
\author{
Dan McGlinn
}
\keyword{internal}

