% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{calc_S_C}
\alias{calc_S_C}
\title{Calculate species richness for a given coverage level.}
\usage{
calc_S_C(x, C_target = NULL, extrapolate = TRUE, interrupt = TRUE)
}
\arguments{
\item{x}{a site by species matrix}

\item{C_target}{target coverage between 0 and 1 (default is NULL). If not
provided then target coverage is computed by \code{\link{calc_C_target}}}

\item{extrapolate}{logical. Defaults to TRUE in which case richness is 
extrapolated to sample sizes larger than observed in the dataset.}

\item{interrupt}{logical. Should the function throw an error when \code{C_target}
exceeds the maximum recommendable coverage?}
}
\value{
a numeric value
}
\description{
Beta_C uses coverage-based rarefaction to standardize beta-diversity.
Specifically, the metric is computed as the Calculates the ratio between gamma and alpha scale IBR curve for a target
gamma-scale sample coverage (i.e. a measure of sample completeness).
}
\examples{
data(tank_comm)
# What is beta_C for a coverage value of 60\%?
calc_S_C(tank_comm, C_target = 0.6)
}
