% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-NCN_curve_functions.R
\name{centroid_accumulate}
\alias{centroid_accumulate}
\title{Internal function used by kNCN_average to compute the k-NCN algorithm
starting with a specified focal sample}
\usage{
centroid_accumulate(
  x,
  focal_sample = 1,
  n = NULL,
  coords = NULL,
  latlong = FALSE
)
}
\arguments{
\item{x}{a mob_in object or a community site x species matrix}

\item{focal_sample}{an integer from 1 to the number of samples of x that 
specifies which sample to start with. Defaults to 1}

\item{n}{the number of samples to accumulate, defaults to NULL in which 
case all samples are accumulated}

\item{coords}{the spatial coordinates of the samples of x}

\item{latlong}{if latitude longitude arguments are supplied}
}
\description{
Internal function used by kNCN_average to compute the k-NCN algorithm
starting with a specified focal sample
}
\keyword{internal}
