% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-NCN_curve_functions.R
\name{compare_curves}
\alias{compare_curves}
\title{Compare all sample-based curves (random, spatiallyconstrained-k-NN, spatially constrained-k-NCN)}
\usage{
compare_curves(x)
}
\arguments{
\item{x}{a mob_in object}
}
\value{
a plot
}
\description{
This is just plotting all curves.
}
