% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-NCN_curve_functions.R
\name{compare_samp_rarefaction}
\alias{compare_samp_rarefaction}
\title{Compare all sample-based curves (random, spatially constrained-k-NN,
 spatially constrained-k-NCN)}
\usage{
compare_samp_rarefaction(x)
}
\arguments{
\item{x}{a mob_in object}
}
\value{
a plot
}
\description{
This is just plotting all curves.
}
\examples{
\donttest{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr, coord_names = c('x', 'y'))
compare_samp_rarefaction(inv_mob_in)
}
}
