% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{deltaS_N}
\alias{deltaS_N}
\title{Difference in S due to N}
\usage{
deltaS_N(comm, ref_dens, inds)
}
\arguments{
\item{comm}{community matrix with plots as rows and species columns.}

\item{ref_dens}{the reference density}

\item{inds}{the number of individuals to sample over}
}
\value{
a two column data.frame containing the number of individuals (inds)
  and the difference in species richness (deltaS)
}
\description{
Difference between the individual and non-spatial sample-based
  rarefaction curves for one group with the evaluation sample size (number of
  individuals) defined by ref_dens, evaluated at specified points (given by
  inds). The rescaling of sampling effort from number of samples to number
  of individuals is accomplished using the mean density of individuals 
  per sample.
}
\details{
Internal function for computing the difference in species richness between 
individual-based and non-spatial sample-based rarefaction curves using an 
analytical approach of stretching (ref_dens > group_dens) or shrinking
(ref_dens < group_dens) the individual-based rarefaction curve to provide the
non-spatial rarefaction result.
}
\author{
Dan McGlinn and Xiao Xiao
}
\keyword{internal}
