% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{effect_agg_continuous}
\alias{effect_agg_continuous}
\title{Auxiliary function for get_delta_stats()
Effect of aggregation when type is "continuous"}
\usage{
effect_agg_continuous(mob_in, sample_rare, group_plots, group_levels,
  group_data, env_levels, corr, n_perm)
}
\description{
Auxiliary function for get_delta_stats()
Effect of aggregation when type is "continuous"
}
\keyword{internal}
