% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{effect_N_discrete}
\alias{effect_N_discrete}
\title{Auxillary function for get_delta_stats()
Effect of N when type is "discrete"}
\usage{
effect_N_discrete(mob_in, group_levels, ref_group, groups, density_stat,
  ind_sample_size, nperm)

effect_N_discrete(mob_in, group_levels, ref_group, groups, density_stat,
  ind_sample_size, nperm)
}
\description{
Auxillary function for get_delta_stats()
Effect of N when type is "discrete"

Auxillary function for get_delta_stats()
Effect of N when type is "discrete"
}
\keyword{internal}

