% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_delta_stats}
\alias{get_delta_stats}
\title{Conduct the MoB tests on drivers of biodiversity across scales.}
\usage{
get_delta_stats(mob_in, group_var, env_var = NULL, ref_group = NULL,
  tests = c("SAD", "N", "agg"), type = "discrete", inds = NULL,
  log_scale = FALSE, min_plots = NULL, density_stat = "mean",
  corr = "spearman", nperm = 1000, overall_p = FALSE)
}
\arguments{
\item{mob_in}{an object of class mob_in created by make_mob_in()}

\item{group_var}{a character string specifying the environmental variable in
mob_in$env used for grouping plots}

\item{env_var}{an optional character string specifying a environmental variable
in mob_in$env which is used in correlation analysis in the continuous case. 
It is not needed if "type" is discrete or group_var is used in correlation.}

\item{ref_group}{a character string used to define the reference group to
which all other groups are compared with when "type" is discrete. It is not
needed when "type" is continuous.}

\item{tests}{specifies which one or more of the three tests ('SAD', N', 'agg') 
are to be performed. Default is to include all three tests.}

\item{type}{"discrete" or "continuous". If "discrete", pair-wise comparisons
are conducted between all other groups and the reference group. If
"continuous", a correlation analysis is conducted between the response
variables and group_var or env_var (if defined).}

\item{inds}{effort size at which the individual-based rarefaction curves are
to be evaluated, and to which the sample-based rarefaction curves are to be
interpolated. It can take three types of values, a single integer, a vector
of intergers, and NULL. If inds = NULL (default), the curves are evaluated
at every possible effort size, from 1 to the total number of individuals
within the group (slow). If inds is a single integer, it is taken as the
number of points at which the curves are evaluated; the positions of the
points are determined by the "log_scale" argument. If inds is a vector of
integers, it is taken as the exact points at which the curves are
evaluated.}

\item{log_scale}{if "inds" is given a single integer, "log_scale" determines
the position of the points. If log_scale is TRUE, the points are equally
spaced on logarithmic scale. If it is FALSE (default), the points are
equally spaced on arithmetic scale.}

\item{min_plot}{minimal number of plots for test 'agg', where plots are
randomized within groups as null test. If it is given a value, all groups
with fewer plots than min_plot are removed for this test. If it is NULL
(default), all groups are kept. Warnings are issued if 1. there is only one
group left and "type" is discrete, or 2. there are less than three groups
left and "type" is continuous, or 3. reference group ("ref_group") is
removed and "type" is discrete. In these three scenarios, the function will
terminate. A different warning is issued if any of the remaining groups
have less than five plots (which have less than 120 permutations), but the 
test will be carried out.}

\item{density_stat}{reference density used in converting number of plots to
numbers of individuals, a step in test "N". It can take one of the
three values: "mean", "max", or "min". If it is "mean", the average
plot-level abundance across plots (all plots when "type" is "continuous,
all plots within the two groups for each pair-wise comparison when "type"
is "discrete") are used. If it is "min" or "max", the minimum/maximum
plot-level density is used.}

\item{corr}{which kind of correlation to use when "type" is "continuous". It
can take two values, "spearman" or "pearson". "spearman" (default) is
generally recommended because the relationship between the response and
"env_var" may not be linear.}

\item{nperm}{number of iterations to run for null tests, defaults to 1000.}

\item{overall_p}{boolean defaults to FALSE specifies if overall across scale 
p-values for the null tests. This should be interpreted with caution because
the overall p-values depend on scales of measurement yet do not explicitly 
reflect significance at any particular scale.}
}
\value{
a "mob_out" object with attributes
}
\description{
There are three tests, on effects of 1. the shape of the SAD, 2.
treatment/group-level density, 3. degree of aggregation. The user can
specificy to conduct one or more of these tests.
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in= make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                           type='discrete', log_scale=TRUE, nperm=20)
}
\author{
Xiao Xiao and Dan McGlinn
}
