% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{get_F_values}
\alias{get_F_values}
\title{Get F statistics from diversity indices and grouping vector}
\usage{
get_F_values(div_dat, permute = FALSE)
}
\description{
Get F statistics from diversity indices and grouping vector
}
\keyword{internal}
