% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{get_group_delta}
\alias{get_group_delta}
\title{Get gamma-scale differences}
\usage{
get_group_delta(
  abund_mat,
  groups,
  index,
  effort,
  extrapolate,
  return_NA,
  rare_thres,
  permute = FALSE
)
}
\description{
Get gamma-scale differences
}
\keyword{internal}
