% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_ind_dens}
\alias{get_ind_dens}
\title{Auxiliary function for get_delta_stats()
Returns the "assumed" density of individuals in 
a plot given whether min, max or mean is used}
\usage{
get_ind_dens(comm, density_stat)
}
\description{
Auxiliary function for get_delta_stats()
Returns the "assumed" density of individuals in 
a plot given whether min, max or mean is used
}
\keyword{internal}
