% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_inds}
\alias{get_inds}
\title{Auxiliary function for get_delta_stats()
Returns a vector of abundances where individual-based rarefaction 
will be performed}
\usage{
get_inds(N_max, inds = NULL, log_scale = FALSE)
}
\description{
Auxiliary function for get_delta_stats()
Returns a vector of abundances where individual-based rarefaction 
will be performed
}
\keyword{internal}
