% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{get_mob_stats}
\alias{get_mob_stats}
\title{Carry out biodiversity metric comparisons between groups.}
\usage{
get_mob_stats(
  mob_in,
  group_var,
  index = c("N", "S", "S_n", "S_PIE"),
  effort_samples = NULL,
  effort_min = 5,
  extrapolate = TRUE,
  return_NA = FALSE,
  rare_thres = 0.05,
  scales = c("alpha", "gamma", "beta"),
  PIE_replace = FALSE,
  C_target_gamma = NA,
  n_perm = 199,
  cl = NULL,
  ci = TRUE,
  ci_cent_stat = "median",
  ci_algo = "boot",
  ci_n_boot = 1000,
  quiet_mode = TRUE,
  ...
)
}
\arguments{
\item{mob_in}{an object of class mob_in created by make_mob_in()}

\item{group_var}{String that specifies which variable in \code{mob_in$env} the
data should be grouped by.}

\item{index}{The calculated biodiversity indices. The options are
\itemize{
   \item \code{N} ... Number of individuals (total abundance)
   \item \code{S} ... Number of species
   \item \code{S_n} ... Rarefied or extrapolated number of species for n individuals
   \item \code{S_C} ... Estimate species richness of a given level of coverage by \code{C_target_gamma}
   \item \code{S_asymp} ... Estimated asymptotic species richness
   \item \code{f_0} ... Estimated number of undetected species
   \item \code{pct_rare} ... The percent of rare species as defined by \code{rare_thres}
   \item \code{PIE} ... Hurlbert's PIE (Probability of Interspecific Encounter)
   \item \code{S_PIE} ... Effective number of species based on PIE

}
  See \emph{Details} for additional information on the biodiversity
  statistics.}

\item{effort_samples}{An integer that specifies the standardized number of
individuals used for the calculation of rarefied species richness at the
alpha-scale. It must be a single integer. The default value of \code{NULL}
will result in the using the minimum number of individuals found across the
samples is used, when this is not smaller than \code{effort_min}.}

\item{effort_min}{The minimum number of individuals considered for the 
calculation of rarefied richness (Default value of 5). Samples with less
individuals then \code{effort_min} are excluded from the analysis with a
warning. Accordingly, when \code{effort_samples} is set by the user it has
to be higher than \code{effort_min}.}

\item{extrapolate}{Boolean which specifies if richness should be extrapolated
when effort is larger than the number of individuals using the chao1
method (Chao 1984, 1987). Defaults to TRUE.}

\item{return_NA}{Boolean in which the rarefaction function returns the
observed S when \code{effort} is larger than the number of individuals. If
set to TRUE then NA is returned. Note that this argument is only relevant
when \code{extrapolate = FALSE}.}

\item{rare_thres}{The threshold that determines how the metric
\code{pct_rare} is computed. It can range from (0, 1] and defaults to 0.05
which specifies that any species with less than or equal to 5% of the total
abundance in a sample is considered rare. It can also be specified as "N/S"
which results in using average abundance as the threshold which McGill
(2011) found to have the best small sample behavior.}

\item{scales}{The scales to compute the diversity indices for:
\itemize{
    \item \code{alpha} ... for each row of the site x species community matrix
    \item \code{gamma} ... for the entire site x species community matrix
    \item \code{beta} ... the ratio of diversity at the \code{gamma} and
                           \code{alpha} scales.
} Defaults to all three scales: \code{c('alpha', 'gamma', 'beta')}}

\item{PIE_replace}{Used for \code{PIE} and \code{SPIE}.  If TRUE, sampling with
replacement is used. Otherwise, sampling without replacement (default).}

\item{C_target_gamma}{When computing coverage based richness (\code{S_C})
then this argument can be used to specify the coverage to be used for the
gamma scale richness estimate. This defaults to \code{NA} in which case the
target cover is computed by \code{\link{calc_C_target}} (i.e., the largest
allowable sample size).}

\item{n_perm}{number of iterations to run for null tests, defaults to 1000.}

\item{cl}{
A cluster object created by \code{\link{makeCluster}},
or an integer to indicate number of child-processes
(integer values are ignored on Windows) for parallel evaluations
(see Details on performance).
It can also be \code{"future"} to use a future backend (see Details),
\code{NULL} (default) refers to sequential evaluation.
}

\item{ci}{boolean, if TRUE then confidence intervals are calculated. Defaults
to TRUE.}

\item{ci_cent_stat}{a string that is either 'mean' or 'median' which
specifies the measure of central tendency. Defaults to 'median'.}

\item{ci_algo}{can be either 'boot' or 'loo' for bootstrap or leave-one-out
methods respectively. Default value is 'boot'.}

\item{quiet_mode}{boolean, defaults to TRUE in which case warnings from 
the diversity calculations are not returned. These are typically generated
when calculating the metric SPIE which is undefined when all species are
singletons.}

\item{...}{additional arguments that can be passed to \code{\link{calc_div}}}
}
\value{
A list of class \code{mob_stats} that contains two objects: 
1) \code{comm_div} a data.frame of each diversity metrics for each group
at each scale specified, and
2) \code{gorup_tests} a data.frame of the average difference between groups
in their diversity metric (\code{D_bar}) with an associated p-value derived 
from the permutation test.
}
\description{
This function can compute a range of biodiversity metrics, their uncertainty, 
and carries out a permutation test to examine if groups differ in a particular
metric more than would be expected due to random chance.
}
\details{
This function is partially a wrapper for the functions
\code{\link{calc_comm_div}} or \code{\link{calc_comm_div_ci}} that makes
group comparisons easier to implement.



See \code{\link{calc_comm_div}} for more details on the biodiversity indices. 

\strong{Group comparison metric and test}

For each metric group comparison the function computes \code{D_bar}: the
average absolute difference between the groups. At the alpha scale the
indices are averaged first before computing \code{D_bar}.

Permutation tests are carried out for testing differences of the biodiversity
statistics among the groups (Legendre & Legendre 1998). This is accomplished
by using \code{D_bar} as the test statistic and random shuffling the group 
label across the samples. The p-value indicates how many of the permutations 
result in a \code{D_bar} as large as the observed \code{D_bar} value.
}
\examples{
# tank community analysis
data(tank_comm)
data(tank_plot_attr)
tank_mob <- make_mob_in(tank_comm, tank_plot_attr)
tank_stats <- get_mob_stats(tank_mob, 'group', index = c('S', 'S_PIE', 'S_C'),
                            n_perm = 5, ci_n_boot = 5)
tank_stats    
}
\references{
Chao, A. 1984. Nonparametric Estimation of the Number of Classes in a
Population. Scandinavian Journal of Statistics 11:265–270.

Chao, A. 1987. Estimating the population size for capture-recapture data with
unequal catchability. Biometrics, 43, 783-791.

Gotelli, N. J., and R. K. Colwell. 2001. Quantifying biodiversity: procedures
and pitfalls in the measurement and comparison of species richness. Ecology
Letters 4:379–391.

Hurlbert, S. H. 1971. The nonconcept of species diversity: a critique and
alternative parameters. Ecology 52:577–586.

Jost, L. 2007. Partitioning diversity into independent alpha and beta
components. Ecology 88:2427–2439.

McGill, B. J. 2011. Species abundance distributions. Pages 105-122 Biological
Diversity: Frontiers in Measurement and Assessment, eds. A.E. Magurran and
B.J. McGill.

Whittaker, R. H. 1972. Evolution and measurement of species diversity. Taxon
21:213–251.
}
