% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_null_comm}
\alias{get_null_comm}
\title{Generate a null community matrix}
\usage{
get_null_comm(comm, tests, groups = NULL)
}
\arguments{
\item{comm}{community matrix with plots as rows and species columns.}

\item{tests}{'SAD', 'N', or 'agg' for a null model that nullifies 
the given component}

\item{groups}{optional argument that is a vector of group ids which specify
which group each site is associated with. If is NULL then all rows of the
community matrix are assumed to be members of the same group}
}
\value{
a site-by-species matrix
}
\description{
## To DO: update this documentation
Two types of permutation can be carried out: 1) 'noagg': each individual of
each species is reassigned a plot randomly which removes any patterns due to
within and between plot spatial aggregation, but maintains species group
abundance and therefore observed group N, and 2) 'swapN': the total number of
individuals in a plot is shuffled and then that many individuals are drawn
randomly from the group specific species-abundance distribution for each plot
which provides a means of removing group differences in the total number of
individuals in a given sample. Note: The 'noagg' algorithm fixes the total number
of individuals for a given species within a group and the 'swapN' algorithm does
not.
}
\details{
Replaces depreciated function `permute_comm`
}
\examples{
S = 3
N = 20
nplots = 4
comm = matrix(rpois(S*nplots, 1),
              ncol=S, nrow=nplots)
comm
groups = rep(1:2, each=2)
groups
get_null_comm(comm, 'noagg')
get_null_comm(comm, 'noagg', groups)
get_null_comm(comm, 'swapN')
get_null_comm(comm, 'swapN', groups)
}
