% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_rand_sad}
\alias{get_rand_sad}
\title{Randomly sample of a relative abundance distribution (RAD)
to produce an expected species abundance distribution (SAD)}
\usage{
get_rand_sad(rad, N)
}
\arguments{
\item{rad}{the relative abundance of each species}

\item{N}{the total number of individuals sampled

Randomly subsampling an RAD with replacement produces an SAD that is of a
similar functional form (Green and Plotkin 2007) but with overall species
richness equal to or less than the relative abundance distribution.

Literature Cited:
Green, J. L., and J. B. Plotkin. 2007. A statistical theory
 for sampling species abundances. Ecology Letters 10:1037-1045.}
}
\description{
Randomly sample of a relative abundance distribution (RAD)
to produce an expected species abundance distribution (SAD)
}
\keyword{internal}
