% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_sad}
\alias{get_sad}
\title{rank abundance distribution}
\usage{
get_sad(indiv_ids)
}
\arguments{
\item{indiv_ids}{a vector of species names were each element of the vector
represents a different individual}
}
\description{
rank abundance distribution
}
\examples{
indivs = c('A', 'A', 'B', 'C', 'C', 'C')
get_sad(indivs)
}

