% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_sample_curves}
\alias{get_sample_curves}
\title{Auxillary function for get_delta_stats()
Obtain the swap curve and/or spatial curve for each group if asked
Directly add attributes to the input "out"}
\usage{
get_sample_curves(mob_in, group_levels, group_data, approved_tests)
}
\description{
Auxillary function for get_delta_stats()
Obtain the swap curve and/or spatial curve for each group if asked
Directly add attributes to the input "out"
}
\keyword{internal}

