% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{groups_panel2}
\alias{groups_panel2}
\title{Panel function for gamma-scale results with confidence intervals}
\usage{
groups_panel2(
  group_dat,
  col,
  ylab = "",
  main = expression(gamma * "-scale"),
  cex.axis = 1.2,
  ...
)
}
\description{
Panel function for gamma-scale results with confidence intervals
}
\keyword{internal}
