% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr-shiny.R
\name{gui}
\alias{gui}
\title{mobr package Graphic User Interface}
\usage{
gui(port = getOption("shiny.port"), host = getOption("shiny.host",
  "127.0.0.1"))
}
\arguments{
\item{port}{char. The TCP port that the application should listen on (see
\code{\link[shiny]{runApp}} for more details).}

\item{host}{char. The IPv4 address that the application should listen on (see
\code{\link[shiny]{runApp}} for more details).}

\item{working.directory}{char. Directory in which the application will run.}
}
\value{
Open a window with a shiny app to use the soar package with an
  user-friendly interface.
}
\description{
User interface of the mobr package.
}
\examples{
\dontrun{
gui()
}

}
