% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{ind_rare_perm}
\alias{ind_rare_perm}
\title{Compute permutation derived individual-based rarefaction curves}
\usage{
ind_rare_perm(abu, n_perm = 100, n_indiv = NULL)
}
\arguments{
\item{abu}{a vector of species abundances}

\item{n_perm}{the number of permutations to average across, defaults to 100}

\item{n_indiv}{the number of individuals to evaluate the rarefaction curve
at. The default behavior is to evaluate it on a log2 interval from 1 to N}
}
\description{
An internal function that can provide an independent derivation of 
the individual rarefaction curve for the purposes of testing the 
performance of the function \code{rarefaction}
}
\keyword{internal}
