% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k-NCN_curve_functions.R
\name{kNCN_average}
\alias{kNCN_average}
\title{Construct spatially constrained sample-based rarefaction (sSBR) curve using the k-Nearest-Centroid-neighbour (k-NCN) algorithm}
\usage{
kNCN_average(x, n = NULL, coords = NULL)
}
\arguments{
\item{x}{a mob_in object or a sites x species matrix}

\item{n}{number of sites to include.}

\item{coords}{spatial coordinates of the samples. If x is a mob_in object, the fuction uses its "spat" table as coords.}
}
\value{
a numeric vector of estimated species richness
}
\description{
This function accumulates samples according their proximity to all previously included samples (their centroid) as opposed
to the proximity to the inital focal sample. This ensures that included samples mutually close to each other and not all over the place.
}
\details{
Internally the function constructs one curve per sample whereby each sample serves as the initial sample once. Finally, the average curve is returned.
}
