% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{make_comm_obj}
\alias{make_comm_obj}
\title{Create a community ('comm') object.}
\usage{
make_comm_obj(x, plot_attr, binary = FALSE)
}
\arguments{
\item{x}{plot (rows) by species (columns) matrix. Values can be species abundances
or presence/absence (1/0).}

\item{plot_attr}{matrix which includes the environmental attributes and spatial 
coordinates of the plots. Environmental attributes are mandatory, while spatial
coordinates are not. If spatial coordinates are provided, the column(s) has to have
names "x" and/or "y".}

\item{binary}{whether the plot by species matrix "comm" is in abundances or presence/absence.}
}
\value{
a "comm" object with four attributes. "comm" is the plot by species matrix. 
 "env" is the environmental attribute matrix, without the spatial coordinates. "spat" 
 contains the spatial coordinates (1-D or 2-D). "tests" specifies whether each of the 
 three tests in the biodiversity analyses is allowed by data.
}
\description{
The 'comm' object will be passed on for analyses of biodiversity across scales.
}
\examples{
 data(inv_sp)
 data(inv_plot_attr)
 inv_comm = make_comm_obj(inv_sp, inv_plot_attr)
 
}

