% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{make_mob_in}
\alias{make_mob_in}
\title{Create the 'mob_in' object.}
\usage{
make_mob_in(comm, plot_attr, binary = FALSE)

make_mob_in(comm, plot_attr, binary = FALSE)
}
\arguments{
\item{comm}{community matrix with plots as rows and species columns.}

\item{plot_attr}{matrix which includes the environmental attributes and 
spatial coordinates of the plots. Environmental attributes are mandatory, 
while spatial coordinates are not. If spatial coordinates are provided, the
column(s) has to have names "x" and/or "y".}

\item{binary}{whether the plot by species matrix "comm" is in abundances or 
presence/absence.}

\item{latlong}{whether the coordinates are in latitude-longitude. If TRUE, it
is assumed that column "x" is longitude and column "y" is latitude, both in 
radians.}

\item{comm}{community matrix with plots as rows and species columns.}

\item{plot_attr}{matrix which includes the environmental attributes and 
spatial coordinates of the plots. Environmental attributes are mandatory, 
while spatial coordinates are not. If spatial coordinates are provided, the
column(s) has to have names "x" and/or "y".}

\item{binary}{whether the plot by species matrix "comm" is in abundances or 
presence/absence.}
}
\value{
a "mob_in" object with four attributes. "comm" is the plot by species 
  matrix. "env" is the environmental attribute matrix, without the spatial 
  coordinates. "spat" contains the spatial coordinates (1-D or 2-D). "tests" 
  specifies whether each of the three tests in the biodiversity analyses is 
  allowed by data.

a "mob_in" object with four attributes. "comm" is the plot by species 
  matrix. "env" is the environmental attribute matrix, without the spatial 
  coordinates. "spat" contains the spatial coordinates (1-D or 2-D). "tests" 
  specifies whether each of the three tests in the biodiversity analyses is 
  allowed by data.
}
\description{
The 'mob_in' object will be passed on for analyses of biodiversity across 
scales.

The 'mob_in' object will be passed on for analyses of biodiversity across 
scales.
}
\examples{
 data(inv_comm)
 data(inv_plot_attr)
 inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
 data(inv_comm)
 data(inv_plot_attr)
 inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
}
\author{
Dan McGlinn and Xiao Xiao

Dan McGlinn and Xiao Xiao
}

