% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{mean_col}
\alias{mean_col}
\title{Compute average color}
\usage{
mean_col(...)
}
\arguments{
\item{...}{Colors to average}
}
\value{
A color string of 9 characters: `"#"` followed by the
        red, blue, green, and alpha values in hexadecimal.
}
\description{
`mean_col()` computes an average color.
}
\details{
Inernal function to compute an average color
by using the quadratic mean of the colors' RGBA values.

Used by \code{\link{plot_rarefaction}}

this function was copied from gridpattern::mean_col (Davis et al. 2024)
}
\examples{
 mean_col("black", "white")
 mean_col(c("black", "white"))
 mean_col("red", "blue")
}
\references{
Davis, T., Mike FC, and ggplot2 authors. 2024. gridpattern. 1.2.1
}
\keyword{internal}
