% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{mob_stats}
\alias{mob_stats}
\title{Calculate sample based and group based biodiversity statistics}
\usage{
mob_stats(mob_in, group_var, plot = T, n_min = 10, nperm = 100)
}
\arguments{
\item{mob_in}{an object of class mob_in created by make_mob_in()}

\item{group_var}{a string that specifies which field in mob_in$env the data
should be grouped by}

\item{nperm}{number of iterations to run for null tests.}
}
\description{
Calculate sample based and group based biodiversity statistics
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
mob_stats(inv_mob_in)
}
\author{
Felix May and Dan McGlinn
}

