% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr-package.R
\docType{package}
\name{mobr}
\alias{mobr}
\alias{mobr-package}
\title{mobr: measurement of biodiversity across scales}
\description{
mobr: measurement of biodiversity across scales
}
\section{About}{


BLAH
}

