% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{mod_sum}
\alias{mod_sum}
\title{Extract coefficients and metrics of fit from model}
\usage{
mod_sum(x, stats = c("betas", "r", "r2", "r2adj", "f", "p"))
}
\arguments{
\item{x}{a fitted model object}

\item{stats}{the statistics to output}
}
\description{
Extract coefficients and metrics of fit from model
}
\keyword{internal}
