% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{overlap_effects}
\alias{overlap_effects}
\title{Plot summary graphics of the effect on species richness}
\usage{
overlap_effects(mob_out, trt_group, display = "raw", prop = FALSE,
  rescale = "max_effort", common_scale = FALSE, xlabel_indiv = TRUE,
  ylim = NULL, log = "", lty = 1, lwd = 3, leg_loc = "topleft",
  col = c("#FFB3B5", "#78D3EC", "#C5C0FE"))
}
\arguments{
\item{mob_out}{a mob_out class object}

\item{display}{the type of graphic to display options include: "raw" or
"stacked" for plots of the raw effect sizes or a stacked area plot of the
absolute value of the effect sizes.}

\item{prop}{boolean if TRUE then proportions are used in the stacked area plot}

\item{rescale}{string that specifies how to rescale number of individuals to
the number of samples. Defaults to 'max_effort' which rescales by 
setting the maximum number of individuals considered in the individual
rarefaction curve to the scale of the maximum number of samples considered
in the spatial rarefaction curve. The other option is 'density_stat' which
rescales individuals to number of samples using the density statistic 
specified when the mob statistics where computed by \code{\link[mobr]{get_delta_stats}}.}

\item{common_scale}{boolean defaults to FALSE. If TRUE then all the effects
are truncated to only be across the same range of number of individuals.}

\item{xlabel_indiv}{boolean if TRUE then the top axis is 
labeled with numbers of individuals}

\item{ylim}{the y limits of the plot.}

\item{log}{a character string which contains \code{"x"} if the x axis
    is to be logarithmic, \code{"y"} if the y axis is to be logarithmic
    and \code{"xy"} or \code{"yx"} if both axes are to be logarithmic.}

\item{lty}{a vector of line types, see \link[graphics]{par}.}

\item{col}{the colors for lines. Three colors can be specified so that each
line can be given its own color with the curves ordered as SAD, N, and
aggregation.}
}
\description{
All three treatment effect sizes due to the SAD, N, or aggregation are
graphed on a single plot. The treatment effect is defined as the treatment difference
in richness due to a particular component of community structure. The overlap
of the three effects is accomplished by rescaling numbers of individuals 
to the number of plots. The effect sizes may be plotted simply in their raw 
form or as stacked area plots.
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                              type='discrete', log_scale=TRUE, n_perm=2)
overlap_effects(inv_mob_out, 'invaded', leg_loc=NA)
overlap_effects(inv_mob_out, 'invaded', display='stacked')
overlap_effects(inv_mob_out, 'invaded', display='stacked', prop=TRUE)
}
