% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot.mob_out}
\alias{plot.mob_out}
\title{Plot mob curves}
\usage{
\method{plot}{mob_out}(mob_out, group = NULL, par_args = NULL,
  same_scale = FALSE)
}
\arguments{
\item{mob_out}{a mob_out class object}

\item{group}{a string that specifies which group to plot. Only required when
comparing multiple discrete groups.}

\item{par_args}{optional argument that sets graphical parameters to set}

\item{same_scale}{if TRUE then all three plots have the same range on the 
y-axis.}
}
\value{
plots the effect of the SAD, the number of individuals, and spatial
 aggregation on the difference in species richness
}
\description{
Plot mob curves
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                           type='discrete', log_scale=TRUE, nperm=20)
plot(inv_mob_out)
plot(inv_mob_out, same_scale=T)
}
\author{
Xiao Xiao
}

