% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot.mobr}
\alias{plot.mobr}
\title{Plot mobr curves}
\usage{
\method{plot}{mobr}(mobr, group = NULL, par_args = NULL,
  same_scale = FALSE)
}
\arguments{
\item{mobr}{a mobr class object}

\item{group}{a string that specifies which group to plot. Only required when
comparing multiple discrete groups.}

\item{par_args}{optional argument that sets graphical parameters to set}

\item{same_scale}{if TRUE then all three plots have the same range on the 
y-axis.}
}
\value{
plots the effect of the SAD, the number of individuals, and spatial
 aggregation on the difference in species richness
}
\description{
Plot mobr curves
}
\examples{
data(inv_sp)
data(inv_plot_attr)
inv_comm = make_comm_obj(inv_sp, inv_plot_attr)
inv_mobr = get_delta_stats(inv_comm, 'group', ref_group='uninvaded',
                           type='discrete', log_scale=TRUE, nperm=20)
plot(inv_mobr)
plot(inv_mobr, same_scale=T)
}

