% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot_rarefaction}
\alias{plot_rarefaction}
\title{Plot rarefaction curves for each treatment group}
\usage{
plot_rarefaction(mob_in, env_var, ref_group, method, pooled = T, col = NULL,
  lwd = 3, log = "", leg_loc = "topleft", ...)
}
\arguments{
\item{mob_in}{a 'mob_in' class object produced by 'make_mob_in'}

\item{env_var}{a string that specifies the column name in mob_in$env that 
specifies the grouping variable.}

\item{ref_group}{a string that specifies the name of the reference group}

\item{method}{either 'indiv', 'samp', or 'spat' for individual, sample, or 
sample spatially explicit based rarefaction respectively}

\item{pooled}{boolean specifying if samples should be pooled at the group
level or not. Defaults to TRUE. This argument only applies when
the individual based rarefaction is used (i.e., method = 'indiv')}

\item{col}{optional vector of colors.}

\item{lwd}{a single value for for line width, see \link[graphics]{par}.}

\item{log}{a character string which contains \code{"x"} if the x axis
    is to be logarithmic, \code{"y"} if the y axis is to be logarithmic
    and \code{"xy"} or \code{"yx"} if both axes are to be logarithmic.}

\item{leg_loc}{the location of the legend. Defaults to 'topleft', 
see \code{\link[graphics]{legend}}. If set to NA then no legend is printed.}

\item{...}{other arguments to provide to \code{\link[mobr]{rarefaction}}}
}
\description{
Plot rarefaction curves for each treatment group
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
# random individual based rarefaction curves
plot_rarefaction(inv_mob_in, 'group', 'uninvaded', 'indiv',
                 pooled=TRUE, leg_loc='bottomright')
plot_rarefaction(inv_mob_in, 'group', 'uninvaded', 'indiv',
                 pooled=FALSE, log='x')
# random sample based rarefaction curves 
plot_rarefaction(inv_mob_in, 'group', 'uninvaded', 'samp',
                 log='xy')
# spatial sample based rarefaction curves 
plot_rarefaction(inv_mob_in, 'group', 'uninvaded', 'spat',
                 log='xy', xy_coords = inv_mob_in$spat)                 
}
