% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot_rarefy}
\alias{plot_rarefy}
\title{Plot different types of rarefaction curves}
\usage{
plot_rarefy(mobr, col = NULL)
}
\arguments{
\item{mobr}{a mobr class object}

\item{col}{}
}
\value{
plots the individual-based, sample-based, and spatially-explict 
sample based rarefaction curves
}
\description{
Plot different types of rarefaction curves
}
\examples{
data(inv_sp)
data(inv_plot_attr)
inv_comm = make_comm_obj(inv_sp, inv_plot_attr)
inv_mobr = get_delta_stats(inv_comm, 'group', ref_group='uninvaded',
                           type='discrete', log_scale=TRUE, nperm=2)
plot_rarefy(inv_mobr)
}

