% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{plot_samples}
\alias{plot_samples}
\title{Boxplots for sample based comparisons}
\usage{
plot_samples(sample_stats, tukey = F, col = NULL)
}
\description{
Boxplots for sample based comparisons
}
\author{
Felix May and Dan McGlinn
}

