% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{plot_SNpie}
\alias{plot_SNpie}
\title{Create 3d plot of richness, abundance, and probability of interspecific 
encounter}
\usage{
plot_SNpie(mob_in, env_var, col = NA)

plot_SNpie(mob_in, env_var, col = NA)
}
\arguments{
\item{mob_in}{a 'mob_in' class object produced by 'make_mob_in'}

\item{env_var}{a string that specifies the column name in mob_in$env that 
specifies the grouping variable.}

\item{col}{optional vector of colors.}

\item{mob_in}{a 'mob_in' class object produced by 'make_mob_in'}

\item{env_var}{a string that specifies the column name in mob_in$env that 
specifies the grouping variable.}

\item{col}{optional vector of colors.}
}
\description{
Create 3d plot of richness, abundance, and probability of interspecific 
encounter

Create 3d plot of richness, abundance, and probability of interspecific 
encounter
}
\examples{
\donttest{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
plot_SNpie(inv_mob_in, 'group')
}
\donttest{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
plot_SNpie(inv_mob_in, 'group')
}
}

