% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot_stack}
\alias{plot_stack}
\title{Plot stacked area curves for richness effects}
\usage{
plot_stack(mob_out, trt_group, stretch = TRUE, common_scale = FALSE)
}
\arguments{
\item{mob_out}{a mob_out class object}

\item{trt_group}{a string that specifies the name of the treatment group}

\item{stretch}{boolean whether or not to rescale individuals
such the maximum number of individuals is at the scale of 
the maximum number of samples}
}
\description{
Plot stacked area curves for richness effects
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                              type='discrete', log_scale=TRUE, nperm=2)
plot_stack(inv_mob_out, 'invaded')
}

