% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{print.mob_in}
\alias{print.mob_in}
\title{Print a shortened verison of the mob_in object}
\usage{
\method{print}{mob_in}(x)

\method{print}{mob_in}(x)
}
\description{
Print a shortened verison of the mob_in object

Print a shortened verison of the mob_in object
}
\keyword{internal}

