% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{print.mob_out}
\alias{print.mob_out}
\title{Print a shortened version of the mob_out object}
\usage{
\method{print}{mob_out}(x)
}
\description{
Print a shortened version of the mob_out object
}
\keyword{internal}
