% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{rarefaction}
\alias{rarefaction}
\title{Rarefaction Species Richness}
\usage{
rarefaction(x, method, effort = NULL, xy_coords = NULL)
}
\arguments{
\item{x}{community data, a matrix-like object or a vector which should either
represent a species-abundance distribution or a plot (rows) by species
(columns) matrix.}

\item{method}{either 'indiv', 'sampl', or 'spat' for individual, sample, or 
sample spatially explicit based rarefaction respectively}

\item{effort}{optional argument to specify what number of individuals or 
number of samples depending on 'method' to compute rarefied richness as. If
not specified all possible values from 1 to the maximum sampling effort are
used}

\item{xy_coords}{the spatial coordinates of the samples only required when 
using the spatial rarefaction method}
}
\value{
A vector of rarefied species richness values
}
\description{
The expected number of species given a particular number of individuals or
samples under and assumption of random sampling with replacement.
}
\details{
The analytical formulas of Cayuela et al. (2015) are used to compute
the random sampling expectation for the individual and sampled based
rarefaction methods. The spatial explicit rarefaction (Chiarucci et al. 2009)
is computed by sampling each plot in the order of their spatial proximity. If
plots have the same distance from the focal plot then one is chosen randomly
to be sampled first. Each plot in the dataset is treated as a starting point
and then the mean of these n possible accumulation curves is computed.
}
\examples{
data(inv_comm)
data(inv_plot_attr)
sad = colSums(inv_comm)
# individual based rarefaction under random sampling
rarefaction(sad, method='indiv', effort=1:10)
# the same results as the above can be generated by supplying the community matrix
rarefaction(inv_comm, method='indiv', effort=1:10)
# sample based rarefaction under random sampling
rarefaction(inv_comm, method='samp')
# sampled based rarefaction under spatially explicit nearest neighbor sampling
rarefaction(inv_comm, method='spat', xy_coords=inv_plot_attr[ , c('x','y')])
}
\author{
Dan McGlinn and Xiao Xiao
}
\references{
Cayuela, L., N.J. Gotelli, & R.K. Colwell (2015) Ecological and 
biogeographic null hypotheses for comparing rarefaction curves. Ecological
Monographs 85:437-454. Appendix A: 
http://esapubs.org/archive/mono/M085/017/appendix-A.php

Chiarucci, A., G. Bacaro, D. Rocchini, C. Ricotta, M. Palmer, & S. Scheiner 
(2009) Spatially constrained rarefaction: incorporating the autocorrelated
structure of biological communities into sample-based rarefaction. Community
Ecology 10:209-214.
}

