% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{samples_panel1}
\alias{samples_panel1}
\title{Panel function for alpha-scale results}
\usage{
samples_panel1(
  sample_dat,
  samples_tests,
  col,
  ylab = "",
  main = expression(alpha * "-scale"),
  cex.axis = 1.2,
  ...
)
}
\description{
Panel function for alpha-scale results
}
\keyword{internal}
