% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mobr.R
\name{site_summary}
\alias{site_summary}
\title{Calcualte site richness and abundance}
\usage{
site_summary(x)
}
\arguments{
\item{x}{a data.frame, a matrix, or a community matrix that has
the class 'indiv_occ', 'site_occ', 'site_by_sp' associated with it}
}
\value{
A data.frame with each sites summary values
}
\description{
This function calculates total abundance (N), richness (S),
and rarified richness (S_rare to the smallest number of individuals
across all sites
}
\details{
rarefied richness is computed under an assumption of random sampling
without replacement (i.e., the hypergeometric distribution)
}
\examples{
nindiv= 100
nsites = 10
stems = data.frame(site = rep(1:nsites, length.out=nindiv),
                   sp = sample(nrow, replace = T))
class(stem_occ) = c('data.frame', 'indiv_occ')
site_summary(stem_occ)
}

