% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{sphere_dist}
\alias{sphere_dist}
\title{Internal function for distance matrix assuming inputs are lat and long 
  on a sphere}
\usage{
sphere_dist(coords)
}
\arguments{
\item{coords}{a matrix with latitudes and longitudes. The longitudes should
be provided in the first column (they are the x-coordinate) and the latitudes
should be provided in the second column (they are the y-coordinate).}
}
\value{
a numeric value
}
\description{
Distance matrix between points on the unit (r = 1) sphere.
}
\author{
Xiao Xiao
}
\keyword{internal}
