% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{stack_effects}
\alias{stack_effects}
\title{Plot stacked area curves for richness effects}
\usage{
stack_effects(mob_out, trt_group, prop = FALSE, stretch = TRUE,
  common_scale = FALSE, label_indv = TRUE, cols = c("#1AB2FF", "#FFBF80",
  "#A6EDFF"))
}
\arguments{
\item{mob_out}{a mob_out class object}

\item{trt_group}{a string that specifies the name of the treatment group}

\item{prop}{boolean if TRUE then propotions are used rather than
raw values}

\item{stretch}{boolean whether or not to rescale individuals
such the maximum number of individuals is at the scale of 
the maximum number of samples}

\item{common_scale}{boolean}

\item{cols}{colors to use for the areas}

\item{label_indiv}{boolean if TRUE then the top axis is 
labeled with numbers of individuals}
}
\description{
Plot stacked area curves for richness effects
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                              type='discrete', log_scale=TRUE, nperm=2)
stack_effects(inv_mob_out, 'invaded')
stack_effects(inv_mob_out, 'invaded', prop=TRUE)
}

