% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_C.R
\name{Chat}
\alias{Chat}
\title{Calculate expected sample coverage C_hat}
\usage{
Chat(x, m)
}
\arguments{
\item{x}{integer vector (species abundances)}

\item{m}{integer. (smaller than observed sample size)}
}
\value{
a numeric value.
}
\description{
Returns expected sample coverage  of a sample x for a smaller than observed sample size m (Chao & Jost, 2012).
This code was copied from INEXT's internal function Chat.Ind() (Hsieh et al 2016).
}
\examples{
\donttest{
library(vegan)
data(BCI)

# What is the expected coverage corresponding to a sample size of 50 at the gamma scale?
Chat(colSums(BCI), 50)
}
}
