% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{deltaS_N}
\alias{deltaS_N}
\title{Difference in S due to N}
\usage{
deltaS_N(comm, ref_dens, inds)
}
\arguments{
\item{comm}{community matrix with plots as rows and species columns.}

\item{ref_dens}{the reference density}

\item{inds}{the number of individuals to sample over}
}
\value{
a two column data.frame containing the number of individuals and the 
  difference in species richness
}
\description{
Difference between the individual and sample-based rarefaction 
  curves for one group with the evaluation sample size (number of individuals)
  defined by ref_dens, evaluated at specified points (given by inds) Output:
  a two-column data frame, with sample size (effort) and deltaS (effect of N)
}
\details{
Internal function for computing the difference in species richness between 
individual-based and sample-based rarefaction curves.
}
\author{
Dan McGlinn and Xiao Xiao
}
\keyword{internal}

