% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{deltaS_N}
\alias{deltaS_N}
\title{Difference in S due to N}
\usage{
deltaS_N(comm_group, ref_dens, inds)
}
\arguments{
\item{comm_group}{the site x species matrix for a specific group}

\item{ref_dens}{the reference density}

\item{inds}{the number of individuals to sample over}
}
\description{
difference between the ind-based rarefaction and the
sample-based rarefaction for one group with the evaluation sample size (number
of individuals) defined by ref_dens, evaluated at specified points (given by
inds) Output: a two-column data frame, with sample size
(effort) and deltaS (effect of N)
}

