% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{effect_agg_continuous}
\alias{effect_agg_continuous}
\title{Auxillary function for get_delta_stats()
Effect of aggregation when type is "continuous"}
\usage{
effect_agg_continuous(mob_in, sample_rare, group_plots, group_levels,
  group_data, env_levels, corr, nperm)

effect_agg_continuous(mob_in, sample_rare, group_plots, group_levels,
  group_data, env_levels, corr, nperm)
}
\description{
Auxillary function for get_delta_stats()
Effect of aggregation when type is "continuous"

Auxillary function for get_delta_stats()
Effect of aggregation when type is "continuous"
}
\keyword{internal}

