% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{effect_agg_discrete}
\alias{effect_agg_discrete}
\title{Auxillary function for get_delta_stats()
Effect of aggregation when type is "discrete"}
\usage{
effect_agg_discrete(mob_in, sample_rare, ref_group, group_plots, group_data,
  group_levels, nperm)

effect_agg_discrete(mob_in, sample_rare, ref_group, group_plots, group_data,
  group_levels, nperm)
}
\description{
Auxillary function for get_delta_stats()
Effect of aggregation when type is "discrete"

Auxillary function for get_delta_stats()
Effect of aggregation when type is "discrete"
}
\keyword{internal}

