% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{effect_SAD_discrete}
\alias{effect_SAD_discrete}
\title{Auxillary function for get_delta_stats()
Effect of SAD when type is "discrete"}
\usage{
effect_SAD_discrete(out, group_sad, group_levels, ref_group, nperm)

effect_SAD_discrete(out, group_sad, group_levels, ref_group, nperm)
}
\description{
Auxillary function for get_delta_stats()
Effect of SAD when type is "discrete"

Auxillary function for get_delta_stats()
Effect of SAD when type is "discrete"
}
\keyword{internal}

