% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{get_mob_stats}
\alias{get_mob_stats}
\title{Calculate sample based and group based biodiversity statistics}
\usage{
get_mob_stats(mob_in, group_var, ref_group = NULL, n_min = 5, nperm = 100)
}
\arguments{
\item{mob_in}{an object of class mob_in created by make_mob_in()}

\item{group_var}{a string that specifies which field in mob_in$env the data
should be grouped by}

\item{ref_group}{a character string used to define the reference group to
which all other groups are compared with when "type" is discrete. It is not
needed when "type" is continuous.}

\item{n_min}{the minimum number of individuals a plot must have to be included
in the estimate of rarefied richness}

\item{nperm}{the number of permutations to use for testing for treatment effects}
}
\value{
a list of class \code{mob_stats} that contains p-values,
 sample-scale (i.e., plot) statistics, and treatment scale statistics
}
\description{
Calculate sample based and group based biodiversity statistics
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_stats = get_mob_stats(inv_mob_in, 'group', nperm=100)
plot(inv_stats, multipanel=TRUE)
}
\author{
Felix May and Dan McGlinn
}

