% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{get_mob_stats}
\alias{get_mob_stats}
\title{Calculate sample based and group based biodiversity statistics.}
\usage{
get_mob_stats(mob_in, group_var, ref_group = NULL, index = c("N", "S",
  "S_rare", "S_asymp", "ENS_PIE"), n_rare_samples = NULL,
  n_rare_groups = NULL, n_perm = 100)
}
\arguments{
\item{mob_in}{an object of class mob_in created by make_mob_in()}

\item{group_var}{A string that specifies which field in \code{mob_in$env} the data
should be grouped by}

\item{ref_group}{The level in \code{group_var} that is used as control group for the
control vs. treatment test at the group level.}

\item{index}{The calculated biodiversity statistics. The options are
\itemize{
   \item \code{N} ... Number of individuals (total abundance)
   \item \code{S} ... Number of species
   \item \code{S_rare} ... Rarefied number of species
   \item \code{S_asymp} ... Estimated asymptotic species richness
   \item \code{PIE} ... Hurlbert's PIE (Probability of Interspecific Encounter)
   \item \code{ENS_PIE} ... Effective number of species based on PIE
}
See \emph{Details} for additional information on the biodiverstiy statistics.}
}
\value{
A list of class \code{mob_stats} that contains sample-scale and 
group-scale biodiversity statistics, as well as the p-values for permutation tests
at both scales.
}
\description{
Calculate sample based and group based biodiversity statistics.
}
\details{
\strong{Rarefied species richness} is the expected number of species, given
a defined number of sampled individuals (Gotelli & Colwell 2001). Rarefied richness
at the sample and group level is calculated for the values provided in
\code{n_rare_samples} and \code{n_rare_groups}, respectively.
When no values are provided the minimum number of individuals of the samples 
or groups is used. \code{S_rare} is calculated using the function \code{\link{rarefaction}}.

\strong{Asymptotic species richness} is calculated using the bias-corrected Chao1
 estimator (Chiu et al. 2014) provided in the function \code{\link[vegan]{estimateR}}.

\strong{PIE} represents the probility that two randomly drawn individuals
belong to the same species. Here we use the definition of Hurlbert (1971), which considers
sampling without replacement. PIE is closely related to the well-known Simpson 
diversity index, but the latter assumes sampling with replacement.

\strong{ENS_PIE} represents the Effective Number of Species derived from PIE.
This corresponds to the the number of equally abundant species, which together result
in the same PIE as the observed community, with - most likely - uneven abundances
(Jost 2006). That means the higher the difference between S and ENS_PIE the more
uneven is the observed community. An intuitive interpretation of ENS_PIE is that
it corresponds to the number of dominant species in the community.

For species richness \code{S} and the Effective Number of Species \code{ENS_PIE}
we also calculate beta-diversity using multiplicative partitioning
(Whittaker 1972, Jost 2007). That means for both indices we estimate beta-diversity as the ratio of group-level
diversity (gamma) divided by sample-level diversity (alpha). 

We used permutation tests for assessing differences of the biodiversity statistics
among the groups (Legendre & Legendre 1998). At the sample level, one-way ANOVA 
(i.e. F-test) is implemented by shuffling treatment group labels across samples.

 At the group-level there is test whether the difference between treatment and control is significantly different from zero. For this purpose we permuted treatment group labels across samples, pooled the groups, and calculated the treatment-control difference for each permutation.
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_stats = get_mob_stats(inv_mob_in, group_var = "group", ref_group = "uninvaded", n_perm = 100, n_rare_samples = c(5,10))
plot(inv_stats)
}
\references{
Chiu, C.-H., Wang, Y.-T., Walther, B.A. & Chao, A. (2014). An improved nonparametric lower bound of species richness via a modified good-turing frequency formula. Biometrics, 70, 671-682.

Gotelli, N.J. & Colwell, R.K. (2001). Quantifying biodiversity: procedures and pitfalls in the measurement and comparison of species richness. Ecology letters, 4, 379-391.

Hurlbert, S.H. (1971). The Nonconcept of Species Diversity: A Critique and Alternative Parameters. Ecology, 52, 577-586.

Jost, L. (2006). Entropy and diversity. Oikos, 113, 363-375.

Jost, L. (2007). Partitioning Diversity into Independent Alpha and Beta Components. Ecology, 88, 2427–2439.

Legendre, P. & Legendre, L.F.J. (1998). Numerical Ecology, Volume 24, 2nd Edition Elsevier, Amsterdam; Boston.

Whittaker, R.H. (1972). Evolution and Measurement of Species Diversity. Taxon, 21, 213-251.
}
\author{
Felix May and Dan McGlinn
}
