% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{get_overall_p}
\alias{get_overall_p}
\title{Auxiliary function for effect_ functions
Compute an overall p-value for one factor in the discrete case
p-value is based on mean squared difference from zero summed across the scales
Method developed by Loosmore and Ford 2006 but algebraic simplifications 
used as developed by Baddeley et al. 2014 Ecological Archives M084-017-A1}
\usage{
get_overall_p(effort, perm, value)
}
\description{
Auxiliary function for effect_ functions
Compute an overall p-value for one factor in the discrete case
p-value is based on mean squared difference from zero summed across the scales
Method developed by Loosmore and Ford 2006 but algebraic simplifications 
used as developed by Baddeley et al. 2014 Ecological Archives M084-017-A1
}
\keyword{internal}
