% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr_boxplots.R
\name{get_samples}
\alias{get_samples}
\title{Generate distribution of sampled community matrices from an original matrix.}
\usage{
get_samples(abund_mat, algo = "boot", n_boot = 199)
}
\arguments{
\item{abund_mat}{Abundance based site-by-species table. Species as columns}

\item{algo}{can be either 'boot' or 'loo' for bootstrap or leave-one-out 
methods respectively.}

\item{n_boot}{how to many boot strapped samples to create, defaults to 199.}
}
\value{
a list of community matrices which are sampled from the original 
input matrix.
}
\description{
The sampled matrices are either bootstrap or leave-one-out samples. The
bootstrap samples represent a random set of the rows sampled with replacement
from the original matrix. The leave-one-out samples are generated by removing
each sample one at a time from the original matrix.
}
\details{
\strong{BIODIVERSITY INDICES}

\strong{N: total community abundance} is the total number of individuals
observed across all species in the sample

\strong{S: species richness} is the observed number of species that occurs at
least once in a sample

\strong{S_n: Rarefied species richness} is the expected number of species,
given a defined number of sampled individuals (n) (Gotelli & Colwell 2001).
Rarefied richness at the alpha-scale is calculated for the values provided in
\code{effort_samples} as long as these values are not smaller than the
user-defined minimum value \code{effort_min}. In this case the minimum value
is used and samples with less individuals are discarded. When no values for
\code{effort_samples} are provided the observed minimum number of individuals
of the samples is used, which is the standard in rarefaction analysis
(Gotelli & Colwell 2001). Because the number of individuals is expected to
scale linearly with sample area or effort, at the gamma-scale the number of
individuals for rarefaction is calculated as the minimum number of samples
within groups multiplied by \code{effort_samples}. For example, when there
are 10
samples within each group, \code{effort_groups} equals \code{10 *
effort_samples}. If n is larger than the number of individuals in sample and
\code{extrapolate = TRUE} then the Chao1 (Chao 1984, 1987) method is used to
extrapolate the rarefaction curve.

\strong{pct_rare: Percent of rare species} Is the ratio of the number of rare
species to the number of observed species x 100 (McGill 2011). Species are
considered rare in a particular sample if they have fewer individuals than
\code{rare_thres * N} where \code{rare_thres} can be set by the user and
\code{N} is the total number of individuals in the sample. The default value
of \code{rare_thres} of 0.05 is arbitrary and was chosen because McGill
(2011) found this metric of rarity performed well and was generally less
correlated with other common metrics of biodiversity. Essentially this metric
attempt to estimate what proportion of the species in the same occur in the
tail of the species abundance distribution and is therefore sensitive to
presence of rare species.

\strong{S_asymp: Asymptotic species richness} is the expected number of
species given complete sampling and here it is calculated using the Chao1
estimator (Chao 1984, Chao 1987) see \code{\link{calc_chao1}}. Note: this
metric is typically highly correlated with S (McGill 2011).

\strong{f_0: Undetected species richness} is the number of undetected species
or the number of species observed 0 times which is an indicator of the degree
of rarity in the community. If there is a greater rarity then f_0 is expected
to increase. This metric is calculated as \code{S_asymp - S}. This metric is
less correlated with S than the raw \code{S_asymp} metric.

\strong{PIE: Probability of intraspecific encounter} represents the
probability that two randomly drawn individuals belong to the same species.
Here we use the definition of Hurlbert (1971), which considers sampling
without replacement. PIE is closely related to the well-known Simpson
diversity index, but the latter assumes sampling with replacement.

\strong{S_PIE: Effective number of species for PIE} represents the effective
number of species derived from the PIE. It is calculated using the asymptotic
estimator for Hill numbers of diversity order 2 (Chao et al. 2014). S_PIE
represents the species richness of a hypothetical community with
equally-abundant species and infinitely many individuals corresponding to the
same value of PIE as the real community. An intuitive interpretation of S_PIE
is that it corresponds to the number of dominant (highly abundant) species in
the species pool.

For species richness \code{S}, rarefied richness \code{S_n}, undetected
richness \code{f_0}, and the Effective Number of Species \code{S_PIE} we also
calculate beta-diversity using multiplicative partitioning (Whittaker 1972,
Jost 2007). That means for these indices we estimate beta-diversity as the
ratio of gamma-diversity (total diversity across all plots) divided by
alpha-diversity (i.e., average plot diversity).
}
\examples{
data(tank)
get_samples(tank)
}
\references{
Chao, A. 1984. Nonparametric Estimation of the Number of Classes in a
Population. Scandinavian Journal of Statistics 11:265–270.

Chao, A. 1987. Estimating the population size for capture-recapture data with
unequal catchability. Biometrics, 43, 783-791.

Gotelli, N. J., and R. K. Colwell. 2001. Quantifying biodiversity: procedures
and pitfalls in the measurement and comparison of species richness. Ecology
Letters 4:379–391.

Hurlbert, S. H. 1971. The nonconcept of species diversity: a critique and
alternative parameters. Ecology 52:577–586.

Jost, L. 2007. Partitioning diversity into independent alpha and beta
components. Ecology 88:2427–2439.

McGill, B. J. 2011. Species abundance distributions. Pages 105-122 Biological
Diversity: Frontiers in Measurement and Assessment, eds. A.E. Magurran and
B.J. McGill.

Whittaker, R. H. 1972. Evolution and measurement of species diversity. Taxon
21:213–251.
}
\seealso{
\code{\link{calc_comm_div_ci}}
}
\author{
Felix May and Dan McGlinn
}
