% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr-package.R
\docType{package}
\name{mobr}
\alias{mobr}
\title{Measurement of Biodiversity in R}
\description{
The primary aim of this package is to provide ecologist's tools
  to examine changes in biodiversity across spatial scales.  Additionally, the
  package provides a method to examine how a factor mediates species richness
  via its effects on different aspects of community structure: total
  abundance, species commonness, and spatial aggregation of conspecifics.
}
