% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot.mob_out}
\alias{plot.mob_out}
\title{Plot mob curves}
\usage{
\method{plot}{mob_out}(mob_out, stat, log = "")
}
\arguments{
\item{mob_out}{a mob_out class object}

\item{log}{a character string which contains \code{"x"} if the x axis
    is to be logarithmic, \code{"y"} if the y axis is to be logarithmic
    and \code{"xy"} or \code{"yx"} if both axes are to be logarithmic.}

\item{trt_group}{a string that specifies the name of the treatment group}

\item{ref_group}{a string that specifies the name of the reference group}

\item{same_scale}{a boolean if TRUE then the y-axis of the rarefaction and 
ddelta S plots are scaled identically across the tested effects}

\item{display}{a string that specifies what graphics to display can be either
'rarefaction', 'delta S', or 'ddelta S' defaults to all three options.}

\item{lwd}{the line width, a single positive number, defaults to \code{3},
see \code{\link[graphics]{par}} for more information.}

\item{leg_loc}{the location of the legend. Defaults to 'topleft', see
\code{\link[graphics]{legend}}. If set to NA then no legend is printed.}

\item{par_args}{optional argument that sets graphical parameters to set}

\item{...}{Other plot arguments}
}
\value{
plots the effect of the SAD, the number of individuals, and spatial
 aggregation on the difference in species richness
}
\description{
Plot mob curves
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                              type='discrete', log_scale=TRUE, n_perm=2)
plot(inv_mob_out, 'b1')
}
\author{
Xiao Xiao and Dan McGlinn
}
