% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot.mob_out}
\alias{plot.mob_out}
\title{Plot mob curves}
\usage{
\method{plot}{mob_out}(mob_out, trt_group, ref_group, same_scale = FALSE,
  display = c("rarefaction", "delta S", "ddelta S"), par_args = NULL)
}
\arguments{
\item{mob_out}{a mob_out class object}

\item{trt_group}{a string that specifies the name of the treatment group}

\item{ref_group}{a string that specifies the name of the reference group}

\item{same_scale}{if TRUE then all three plots have the same range on the 
y-axis.}

\item{display}{argument specifies what graphics to display can be either
'rarefaction', 'delta S', or 'ddelta S' defaults to all three options.}

\item{par_args}{optional argument that sets graphical parameters to set}
}
\value{
plots the effect of the SAD, the number of individuals, and spatial
 aggregation on the difference in species richness
}
\description{
Plot mob curves
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                              type='discrete', log_scale=TRUE, nperm=2)
plot(inv_mob_out, 'invaded', 'uninvaded', display='rarefaction')
plot(inv_mob_out, 'invaded', 'uninvaded', display='delta S')
plot(inv_mob_out, 'invaded', 'uninvaded', display='ddelta S')
}
\author{
Xiao Xiao and Dan McGlinn
}

