% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R, R/mobr_portal_branch_version.R
\name{plot_abu}
\alias{plot_abu}
\title{Plot distributions of species abundance}
\usage{
plot_abu(mob_in, env_var, type = c("sad", "rad"), pooled = FALSE,
  col = NA, log = "", leg_loc = "topleft")

plot_abu(mob_in, env_var, type = c("sad", "rad"), pooled = FALSE,
  col = NA, log = "", leg_loc = "topleft")
}
\arguments{
\item{mob_in}{a 'mob_in' class object produced by 'make_mob_in'}

\item{env_var}{a string that specifies the column name in mob_in$env that 
specifies the grouping variable.}

\item{type}{either 'sad' or 'rad' for species abundance vs rank abundance distribution}

\item{pooled}{boolean specifying if abundances should be pooled at the group level or
not}

\item{col}{optional vector of colors.}

\item{log}{specify which axes to apply log transformations to.}

\item{mob_in}{a 'mob_in' class object produced by 'make_mob_in'}

\item{env_var}{a string that specifies the column name in mob_in$env that 
specifies the grouping variable.}

\item{type}{either 'sad' or 'rad' for species abundance vs rank abundance distribution}

\item{pooled}{boolean specifying if abundances should be pooled at the group level or
not}

\item{col}{optional vector of colors.}

\item{log}{specify which axes to apply log transformations to.}
}
\description{
Plot distributions of species abundance

Plot distributions of species abundance
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
plot_abu(inv_mob_in, 'group', 'sad', pooled=F, log='x')
plot_abu(inv_mob_in, 'group', 'rad', pooled=T, log='x')
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
plot_abu(inv_mob_in, 'group', 'sad', pooled=F, log='x')
plot_abu(inv_mob_in, 'group', 'rad', pooled=T, log='x')
}

