% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{plot_rarefy}
\alias{plot_rarefy}
\title{Plot different types of rarefaction curves}
\usage{
plot_rarefy(mob_out, col = NULL, rare_type = c("SAD", "N", "agg"),
  leg_loc = "topleft", par_args = NULL)
}
\arguments{
\item{mob_out}{a mob_out class object}

\item{col}{optional vector of colors to use}

\item{rare_type}{either 'SAD', 'N', or 'agg', defaults to all three types}

\item{leg_loc}{string that specifies location of the legend}

\item{par_args}{optional argument that sets graphical parameters to set}
}
\value{
plots the individual-based, sample-based, and spatially-explict 
sample based rarefaction curves
}
\description{
Plot different types of rarefaction curves
}
\examples{
data(inv_comm)
data(inv_plot_attr)
inv_mob_in = make_mob_in(inv_comm, inv_plot_attr)
inv_mob_out = get_delta_stats(inv_mob_in, 'group', ref_group='uninvaded',
                           type='discrete', log_scale=TRUE, nperm=2)
plot_rarefy(inv_mob_out)
}
\author{
Xiao Xiao and Dan McGlinn
}

