% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobr.R
\name{sphere_dist}
\alias{sphere_dist}
\title{Internal function for distance matrix assuming inputs are lat and long 
  on a sphere}
\usage{
sphere_dist(long, lat)
}
\arguments{
\item{long}{a vector of longitudes}

\item{lat}{a vector of latitudes, the same length of long}
}
\value{
a numeric value
}
\description{
Distance matrix between points on the unit (r = 1) sphere.
}
\author{
Xiao Xiao
}
\keyword{internal}
